refactor: replace bio and content parsing with UiContentEntitiesRenderer component
