+++ app/components/tweet/TweetView.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { ref, computed } from 'vue';
+import { ref } from 'vue';
 import Avatar from '~/components/ui/Avatar.vue';
 import type { Tweet } from '~~/shared/types/tweets';
 import TweetMedia from './TweetMedia.vue';
@@ -9,7 +9,6 @@ interface Props {
 }
 const props = defineProps<Props>();
 
-type Segment = { type: 'text' | 'mention' | 'hashtag'; text: string; href?: string };
 const tweet = ref(props.tweet);
 
 // Update local tweet state when like/unlike succeeds
@@ -42,62 +41,6 @@ const onUndoRetweetSuccess = () => {
     tweet.value.retweetCount = next < 0 ? 0 : next;
   }
 };
-
-const contentSegments = computed<Segment[]>(() => {
-  const segments: Segment[] = [];
-  const content = tweet.value.content || '';
-  const { entities } = tweet.value;
-  if (!entities || (!entities.mentions?.length && !entities.hashtags?.length)) {
-    return [{ type: 'text', text: content }];
-  }
-
-  type Range = {
-    start: number;
-    end: number;
-    type: 'mention' | 'hashtag';
-    text: string;
-    href: string;
-  };
-  const ranges: Range[] = [];
-
-  for (const m of entities.mentions || []) {
-    const start = m.startPosition;
-    const text = `@${m.username} `;
-    ranges.push({
-      start,
-      end: start + text.length,
-      type: 'mention',
-      text,
-      href: `/@${m.username}`,
-    });
-  }
-  for (const h of entities.hashtags || []) {
-    const start = h.startPosition;
-    const text = `#${h.hashtag} `;
-    ranges.push({
-      start,
-      end: start + text.length,
-      type: 'hashtag',
-      text,
-      href: `/hashtag/${h.hashtag}`,
-    });
-  }
-
-  ranges.sort((a, b) => a.start - b.start);
-
-  let cursor = 0;
-  for (const r of ranges) {
-    if (r.start > cursor) {
-      segments.push({ type: 'text', text: content.slice(cursor, r.start) });
-    }
-    segments.push({ type: r.type, text: r.text, href: r.href });
-    cursor = r.end;
-  }
-  if (cursor < content.length) {
-    segments.push({ type: 'text', text: content.slice(cursor) });
-  }
-  return segments;
-});
 </script>
 
 <template>
@@ -131,14 +74,7 @@ const contentSegments = computed<Segment[]>(() => {
     </div>
     <div class="border-b-border mt-3 border-b-1 pb-3">
       <p class="mt-1 leading-relaxed break-words whitespace-pre-wrap">
-        <template v-for="(seg, i) in contentSegments" :key="i">
-          <span v-if="seg.type === 'text'" class="inline">
-            {{ seg.text }}
-          </span>
-          <NuxtLink v-else :to="seg.href" class="text-primary inline font-medium hover:underline">
-            {{ seg.text }}
-          </NuxtLink>
-        </template>
+        <UiContentEntitiesRenderer :content="tweet.content" :entities="tweet.entities" />
       </p>
       <TweetMedia :media="tweet.media" />
 
@@ -173,14 +109,7 @@ const contentSegments = computed<Segment[]>(() => {
 
       <!-- Content -->
       <p class="mt-1 leading-relaxed break-words whitespace-pre-wrap">
-        <template v-for="(seg, i) in contentSegments" :key="i">
-          <span v-if="seg.type === 'text'" class="inline">
-            {{ seg.text }}
-          </span>
-          <NuxtLink v-else :to="seg.href" class="text-primary inline font-medium hover:underline">
-            {{ seg.text }}
-          </NuxtLink>
-        </template>
+        <UiContentEntitiesRenderer :content="tweet.content" :entities="tweet.entities" />
       </p>
 
       <!-- Media (single image basic layout) -->
