+++ app/components/ui/UserMetadata.vue
@@ -3,7 +3,6 @@ import Avatar from '@/components/ui/Avatar.vue';
 import { useQuery } from '@tanstack/vue-query';
 import FollowToggleButton from './FollowToggleButton.vue';
 import { profileTabsService } from '~/services/profile/profileTabsService';
-import { parseContentEntities } from '~/utils/contentEntityParser';
 const props = defineProps<{
   username: string;
 }>();
@@ -28,17 +27,6 @@ const unfollowUser = () => {
 const unblockUser = () => {
   emit('unblock');
 };
-const parsedBioTokens = computed(() =>
-  user.value
-    ? parseContentEntities(
-        user.value.bio,
-        user.value.bioEntities || {
-          mentions: [],
-          hashtags: [],
-        },
-      )
-    : [],
-);
 </script>
 
 <template>
@@ -69,34 +57,15 @@ const parsedBioTokens = computed(() =>
       </NuxtLink>
     </div>
     <p v-if="user.bio" class="text-sm">
-      <template v-for="token in parsedBioTokens" :key="token.key">
-        <span v-if="token.type === 'text'" :key="token.key">
-          {{ token.display }}
-        </span>
-        <NuxtLink
-          v-else-if="token.type === 'mention'"
-          :to="`/profile/${token.value}`"
-          class="text-primary hover:underline"
-        >
-          {{ token.display }}
-        </NuxtLink>
-        <NuxtLink
-          v-else-if="token.type === 'hashtag'"
-          :to="`/hashtag/${token.value}`"
-          class="text-primary hover:underline"
-        >
-          {{ token.display }}
-        </NuxtLink>
-        <a
-          v-else-if="token.type === 'link'"
-          :href="token.value"
-          target="_blank"
-          rel="noopener noreferrer"
-          class="text-primary hover:underline"
-        >
-          {{ token.display }}
-        </a>
-      </template>
+      <UiContentEntitiesRenderer
+        :content="user.bio"
+        :entities="
+          user.bioEntities ?? {
+            mentions: [],
+            hashtags: [],
+          }
+        "
+      />
     </p>
     <div class="flex gap-5">
       <span class="text-sm font-medium">
